move license overrides closer to module
