+++ script/licenses/license-overrides.ts
